roll back to more stable version of newpipe extractor
