+++ app/build.gradle
@@ -54,7 +54,7 @@ dependencies {
         exclude module: 'support-annotations'
     }
 
-    implementation 'com.github.TeamNewPipe:NewPipeExtractor:0501a2f543'
+    implementation 'com.github.TeamNewPipe:NewPipeExtractor:bf1c771'
 
     testImplementation 'junit:junit:4.12'
     testImplementation 'org.mockito:mockito-core:1.10.19'
